+++ tsconfig.json
@@ -4,7 +4,7 @@
   "exclude": ["node_modules/*"],
   "compilerOptions": {
     "target": "es2020",
-    "module": "es2020",
+    "module": "es2022",
     "types": ["vite/client"],
     "sourceMap": true,
     "baseUrl": "./",
